add more test coverage

+++ internal/app/auth/dependencies.go
@@ -3,7 +3,7 @@ package auth
 type Repository interface {
 	GetUserByEmail(email string) (*User, error)
 	CreateUser(email, password string) error
-	CreateSession(userID int64) (*Session, error)
+	CreateSession(userID int) (*Session, error)
 	GetSessionByToken(token string) (*Session, error)
 	GetAllUsers() ([]*User, error)
 }
